+++ src/main/java/org/bytedeco/javacpp/FloatPointer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2011-2013 Samuel Audet
+ * Copyright (C) 2011-2016 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -102,7 +102,7 @@ public class FloatPointer extends Pointer {
      * @param f the {@code float} value to copy
      * @return this
      */
-    public native FloatPointer put(int i, float f);
+    public native FloatPointer put(long i, float f);
 
     /** @return {@code get(array, 0, array.length)} */
     public FloatPointer get(float[] array) { return get(array, 0, array.length); }
